add Selectionbar component
